requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-WTF==1.1.1
python-dotenv==1.0.0
Pillow==10.0.0
opencv-python==4.8.0.74
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
ultralytics==8.0.196
werkzeug==2.3.7
gunicorn==21.2.0
packaging==23.1
redis==4.6.0
celery==5.3.1
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3